#2923: canabolize 'support-vector' into 'analysis'
