+++ src/jsx/import/redux/container/analysis-layout.jsx
@@ -11,19 +11,23 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import AnalysisLayout from '../../layout/analysis.jsx';
+import { setLayout, setGotoResultsButton, setSvButton } from '../action/page.jsx';
+import setCurrentResult from '../action/current-result.jsx';
 
 // transforms redux state tree to react properties
 const mapStateToProps = (state) => {
     var contentType = false;
-    var displayGotoResults = false;
+    var gotoResultsbtn = false;
+    var submitBtn = false;
 
     if (
         state &&
         state.page &&
-        state.page.button &&
-        !!state.page.button.goto_results
+        state.page.button
     ) {
-            var displayGotoResults = state.page.button.goto_results;
+        const button = state.page.button;
+        var submitBtn = !!button.submit_analysis ? button.submit_analysis : null;
+        var gotoResultsBtn = !!button.goto_results ? button.goto_results : null;
     }
 
     if (
@@ -31,24 +35,35 @@ const mapStateToProps = (state) => {
         state.page &&
         !!state.page.content_type
     ) {
-            var contentType = state.page.content_type;
+        var contentType = state.page.content_type;
     }
 
   // return redux to state
     return {
         page: {
             button: {
-                goto_results: displayGotoResults
+                goto_results: gotoResultBbtn,
+                submit_analysis: submitBtn
             },
             content_type: contentType
         }
     }
 }
 
+// wraps each function of the object to be dispatch callable
+const mapDispatchToProps = (dispatch) => {
+    return {
+        dispatchLayout: dispatch.bind(setLayout),
+        dispatchSvButton: dispatch.bind(setSvButton),
+        dispatchCurrentResult: dispatch.bind(setCurrentResult),
+        dispatchGotoResultsButton: dispatch.bind(setGotoResultsButton)
+    }
+}
+
 // pass selected properties from redux state tree to component
 const AnalysisLayoutState = connect(
     mapStateToProps,
-    null
+    mapDispatchToProps
 )(AnalysisLayout)
 
 // indicate which class can be exported, and instantiated via 'require'
